// services/cCREs-service/graphqlservice/src/postgres/types.ts
export type GenomicRange = {
    chromosome: string;
    start: number;
    end: number;
};

export type BiosampleEntry = {
    biosample_name: string;
    [key: string]: string | null | undefined;
};

export type RDHSEntry = {
    accession: string;
    chromosome: string;
    start: number;
    stop: number;
    assembly?: string;
};
// ...
export type RDHSParemeters = {
    assembly: string;
    accession_prefix?: string[];
    accession?: string[];
    coordinates?: GenomicRange[];
    limit?: number;
    activeInBiosamples?: string[];
};
// ...
export type CCREMaxZParameters = {
    assembly: string;
    accession: string[];
    assay: string;
};

export type CCREMaxZEntry = {
    accession: string;
    score: number;
};